app/pyproject.toml
[project]
name = "ai-learning-path"
version = "0.1.0"
description = "AI-powered learning path generator for Chinese high school mathematics"
authors = [
    {name = "Devin",email = "devin@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faiss-cpu (>=1.9.0.post1,<2.0.0)",
    "sentence-transformers (>=3.3.1,<4.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
